+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 
 - Added build badge for builds that are new
+- Added description tags for Amplitude `[A]` and Resonanace `[R]`
 
 ### Fixed